Removed Input class -> ported to OpenGL module as InputManager

+++ Engine/src/GUI/GuiLayer.cpp
@@ -39,7 +39,7 @@ namespace SW
 	{
 		ImGuiIO& io      = ImGui::GetIO();
 		Application* app = Application::Get();
-		io.DisplaySize   = ImVec2(app->GetWindow()->GetWidth(), app->GetWindow()->GetHeight());
+		io.DisplaySize   = ImVec2((float)app->GetWindow()->GetWidth(), (float)app->GetWindow()->GetHeight());
 		io.DeltaTime     = dt;
 	}
